+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Changed
 
 - Now the exports are named functions (useful for debugging).
+- add-months: proto.js injects the setters `addMonths` and `addUTCMonths`.
+  This goes more according to how the rest of the `Date.prototype` methods work. Because the 1.0 is a version without diffusion, the change is made in the minor version.
 
 ### Fixed
 
@@ -17,7 +19,7 @@
   index.d.ts had an incorrect `export`, now the CJS build is done without Rollup, which does not work with the `export =` syntax of TS.
   Also, the tests failed after a "clean" when not finding proto.js; now they are generated during the CJS build.
 
-- escape-regex-str: Bad written example showing incorrect result.
+- escape-regex-str: Bad written example, showing incorrect result.
 - Issues with markdown linter.
 
 ## \[1.0.0] - 2018-11-09